+++ examples/config/compose.toml
@@ -9,7 +9,7 @@ addrs = ["0.0.0.0:7000", "[::]:7000"]
 [[backends]]
 address = "backend-a:9000"
 http_version = "http11"
-# health_check = { type = "http", path = "/", expected_status = 200, interval_secs = 2, timeout_secs = 2 }
+health_check = { type = "http", path = "/", expected_status = 200, interval_secs = 2, timeout_secs = 2 }
 
 [[backends]]
 address = "backend-b:9000"
