+++ huginn-proxy-lib/tests/backend/health_check/mod.rs
@@ -1,3 +1,4 @@
+mod check_http;
 mod counter_and_tcp_probe;
 pub mod health;
 pub mod registry;
